Give new UI components their own special place
