+++ TabletopTweaks/NewComponents/AbilitySpecific/QuickStudyComponent.cs
@@ -182,101 +182,7 @@ namespace TabletopTweaks.NewComponents.AbilitySpecific {
             }
         }
 
-        [HarmonyPatch(typeof(ActionBarSpontaneousConvertedSlot), "Set", new Type[] { typeof(UnitEntityData), typeof(AbilityData) })]
-        static class ActionBarSpontaneousConvertedSlot_Set_QuickStudy_Patch {
-            static bool Prefix(ActionBarSpontaneousConvertedSlot __instance, UnitEntityData selected, AbilityData spell) {
-                var pseudoActivatableComponent = spell.Blueprint.GetComponent<PseudoActivatable>();
-                if (pseudoActivatableComponent != null) {
-                    __instance.Selected = selected;
-                    if (selected == null) {
-                        return true;
-                    }
-                    __instance.MechanicSlot = new MechanicActionBarSlotPseudoActivatableAbility {
-                        Spell = spell,
-                        Unit = selected,
-                        BuffToWatch = pseudoActivatableComponent.BuffToWatch
-                    };
-                    __instance.MechanicSlot.SetSlot(__instance);
-                    return false;
-                } else if (spell.Blueprint.GetComponent<QuickStudyComponent>()) {
-                    __instance.Selected = selected;
-                    if (selected == null) {
-                        return true;
-                    }
-                    __instance.MechanicSlot = new MechanicActionBarSlotQuickStudy {
-                        Spell = spell,
-                        Unit = selected
-                    };
-                    __instance.MechanicSlot.SetSlot(__instance);
-                    return false;
-                }
-                return true;
-            }
-        }
-
-        [HarmonyPatch(typeof(ActionBarSlotVM), "OnShowConvertRequest")]
-        static class ActionBarSlotVM_OnShowConvertRequest_QuickStudy_Patch {
-            static bool Prefix(ActionBarSlotVM __instance) {
-                if (__instance.ConvertedVm.Value != null && !__instance.ConvertedVm.Value.IsDisposed) {
-                    __instance.CloseConvert();
-                    return false;
-                }
-                if (__instance.m_Conversion.Count == 0) {
-                    return false;
-                }
-                __instance.ConvertedVm.Value = new ActionBarConvertedVM(__instance.m_Conversion.Select(abilityData => {
-                    var pseudoActivatable = abilityData.Blueprint.GetComponent<PseudoActivatable>();
-                    if (pseudoActivatable != null) {
-                        return new MechanicActionBarSlotPseudoActivatableAbility {
-                            Spell = abilityData,
-                            Unit = __instance.MechanicActionBarSlot.Unit,
-                            BuffToWatch = pseudoActivatable.BuffToWatch
-                        };
-                    } else if (abilityData.Blueprint.GetComponent<QuickStudyComponent>()) {
-                        return new MechanicActionBarSlotQuickStudy {
-                            Spell = abilityData,
-                            Unit = __instance.MechanicActionBarSlot.Unit
-                        };
-                    } else {
-                        return new MechanicActionBarSlotSpontaneusConvertedSpell {
-                            Spell = abilityData,
-                            Unit = __instance.MechanicActionBarSlot.Unit
-                        };
-                    }
-
-                }).ToList(), new Action(__instance.CloseConvert));
-                return false;
-            }
-        }
-
-        class MechanicActionBarSlotQuickStudy : MechanicActionBarSlotSpontaneusConvertedSpell {
-
-            public override string GetTitle() {
-                return $"{Spell.Name} - {Spell.m_ConvertedFrom.Name}";
-            }
-
-            public override TooltipBaseTemplate GetTooltipTemplate() {
-                return new TooltipTemplateQuickStudy(Spell);
-            }
-
-            public override Sprite GetIcon() {
-                return Spell.m_ConvertedFrom.Icon;
-            }
-
-            public override Sprite GetForeIcon() {
-                return null;
-            }
-
-            public override Sprite GetDecorationSprite() {
-                return UIUtility.GetDecorationBorderByIndex(Spell.m_ConvertedFrom.DecorationBorderNumber);
-            }
-
-            public override Color GetDecorationColor() {
-                return UIUtility.GetDecorationColorByIndex(Spell.m_ConvertedFrom.DecorationColorNumber);
-            }
-        }
-
-        class TooltipTemplateQuickStudy : TooltipTemplateAbility {
+        public class TooltipTemplateQuickStudy : TooltipTemplateAbility {
 
             public TooltipTemplateQuickStudy(AbilityData abilityData) : base(abilityData) {
             }
